refactor(subprocess): extract internal subprocess execution to dedicated module
- Move subprocess execution logic from cuprum/sh.py to new module cuprum/_subprocess_execution.py
- Encapsulate subprocess spawning, stream handling, and timeout management
- Simplify cuprum/sh.py by importing execution functions from the new module
- Improve code organization and reduce circular imports
- Adjust related context and observation handling to use new subprocess execution flow
- Update tests and docs for consistent usage and error messages
