+++ cuprum/sh.py
@@ -10,8 +10,6 @@ from __future__ import annotations
 import asyncio
 import collections.abc as cabc
 import dataclasses as dc
-import sys
-import time
 import typing as typ
 from pathlib import Path
 
@@ -29,10 +27,10 @@ from cuprum._pipeline_internals import (
     _StageObservation,
 )
 from cuprum._pipeline_streams import _prepare_pipeline_config
-from cuprum._process_lifecycle import _merge_env, _terminate_process
-from cuprum._streams import (
-    _consume_stream,
-    _StreamConfig,
+from cuprum._subprocess_execution import (
+    _execute_subprocess,
+    _resolve_timeout,
+    _SubprocessExecution,
 )
 from cuprum.catalogue import (
     DEFAULT_CATALOGUE,
@@ -40,7 +38,6 @@ from cuprum.catalogue import (
     ProjectSettings,
 )
 from cuprum.catalogue import UnknownProgramError as UnknownProgramError
-from cuprum.context import current_context as _current_context
 from cuprum.context import observe as observe
 from cuprum.context import scoped as scoped
 
@@ -226,217 +223,6 @@ class TimeoutExpired(TimeoutError):  # noqa: N818  # match subprocess.TimeoutExp
         return self.output
 
 
-def _resolve_timeout(
-    *,
-    timeout: float | None,
-    context: ExecutionContext | None,
-) -> float | None:
-    """Resolve the effective timeout from explicit, context, and scoped values."""
-    if timeout is not None:
-        return timeout
-    if context is not None and context.timeout is not None:
-        return context.timeout
-    return _current_context().timeout
-
-
-async def _wait_for_exit_code(
-    process: asyncio.subprocess.Process,
-    ctx: ExecutionContext,
-    *,
-    timeout: float | None = None,
-    consumers: tuple[asyncio.Task[typ.Any], ...] = (),
-) -> tuple[int, float]:
-    """Wait for a subprocess, handling cancellation and capturing exit time."""
-    try:
-        if timeout is None:
-            exit_code = await process.wait()
-        else:
-            exit_code = await asyncio.wait_for(process.wait(), timeout)
-    except TimeoutError:
-        await _terminate_process(process, ctx.cancel_grace)
-        if consumers:
-            await asyncio.gather(*consumers, return_exceptions=True)
-        raise
-    except asyncio.CancelledError:
-        await _terminate_process(process, ctx.cancel_grace)
-        if consumers:
-            await asyncio.gather(*consumers, return_exceptions=True)
-        raise
-    else:
-        exited_at = time.perf_counter()
-        return exit_code, exited_at
-
-
-@dc.dataclass(frozen=True, slots=True)
-class _SubprocessExecution:
-    cmd: SafeCmd
-    ctx: ExecutionContext
-    capture: bool
-    echo: bool
-    timeout: float | None
-    observation: _StageObservation
-
-
-class _SubprocessTimeoutError(Exception):
-    """Internal wrapper for subprocess timeouts with captured output."""
-
-    def __init__(
-        self,
-        *,
-        timeout: float,
-        stdout: str | None,
-        stderr: str | None,
-        exited_at: float,
-    ) -> None:
-        super().__init__(f"Execution exceeded {timeout}s timeout")
-        self.timeout = timeout
-        self.stdout = stdout
-        self.stderr = stderr
-        self.exited_at = exited_at
-
-
-async def _spawn_subprocess(
-    execution: _SubprocessExecution,
-) -> asyncio.subprocess.Process:
-    return await asyncio.create_subprocess_exec(
-        *execution.cmd.argv_with_program,
-        stdout=(
-            asyncio.subprocess.PIPE
-            if execution.capture or execution.echo
-            else asyncio.subprocess.DEVNULL
-        ),
-        stderr=(
-            asyncio.subprocess.PIPE
-            if execution.capture or execution.echo
-            else asyncio.subprocess.DEVNULL
-        ),
-        env=_merge_env(execution.ctx.env),
-        cwd=(str(execution.ctx.cwd) if execution.ctx.cwd is not None else None),
-    )
-
-
-def _create_stream_callback(
-    observation: _StageObservation,
-    event_type: typ.Literal["stdout", "stderr"],
-    pid: int | None,
-) -> cabc.Callable[[str], None] | None:
-    """Create a callback for emitting stream line events, or None if no hooks."""
-    if not observation.hooks.observe_hooks:
-        return None
-    return lambda line: observation.emit(event_type, _EventDetails(pid=pid, line=line))
-
-
-def _spawn_stream_consumers(
-    process: asyncio.subprocess.Process,
-    execution: _SubprocessExecution,
-    stream_config: _StreamConfig,
-    *,
-    pid: int | None,
-) -> tuple[asyncio.Task[str | None], asyncio.Task[str | None]]:
-    """Spawn stdout and stderr stream consumer tasks."""
-    stdout_on_line = _create_stream_callback(execution.observation, "stdout", pid)
-    stderr_on_line = _create_stream_callback(execution.observation, "stderr", pid)
-    stderr_config = dc.replace(
-        stream_config,
-        sink=(
-            execution.ctx.stderr_sink
-            if execution.ctx.stderr_sink is not None
-            else sys.stderr
-        ),
-    )
-    return (
-        asyncio.create_task(
-            _consume_stream(
-                process.stdout,
-                stream_config,
-                on_line=stdout_on_line,
-            ),
-        ),
-        asyncio.create_task(
-            _consume_stream(
-                process.stderr,
-                stderr_config,
-                on_line=stderr_on_line,
-            ),
-        ),
-    )
-
-
-async def _run_subprocess_with_streams(
-    process: asyncio.subprocess.Process,
-    execution: _SubprocessExecution,
-    *,
-    pid: int | None,
-    timeout: float | None,
-) -> tuple[int, float, str | None, str | None]:
-    stream_config = _StreamConfig(
-        capture_output=execution.capture,
-        echo_output=execution.echo,
-        sink=(
-            execution.ctx.stdout_sink
-            if execution.ctx.stdout_sink is not None
-            else sys.stdout
-        ),
-        encoding=execution.ctx.encoding,
-        errors=execution.ctx.errors,
-    )
-    consumers = _spawn_stream_consumers(process, execution, stream_config, pid=pid)
-    try:
-        exit_code, exited_at = await _wait_for_exit_code(
-            process,
-            execution.ctx,
-            timeout=timeout,
-            consumers=consumers,
-        )
-    except TimeoutError as exc:
-        stdout_text, stderr_text = await asyncio.gather(*consumers)
-        if timeout is None:
-            msg = "TimeoutError without a configured timeout"
-            raise RuntimeError(msg) from exc
-        raise _SubprocessTimeoutError(
-            timeout=timeout,
-            stdout=stdout_text,
-            stderr=stderr_text,
-            exited_at=time.perf_counter(),
-        ) from exc
-    stdout_text, stderr_text = await asyncio.gather(*consumers)
-    return exit_code, exited_at, stdout_text, stderr_text
-
-
-def _get_exit_code(process: asyncio.subprocess.Process) -> int:
-    """Return the process exit code, defaulting to -1 if unavailable."""
-    return process.returncode if process.returncode is not None else -1
-
-
-def _emit_exit_event(  # noqa: PLR0913
-    observation: _StageObservation,
-    *,
-    pid: int | None,
-    exit_code: int,
-    started_at: float,
-    exited_at: float,
-) -> None:
-    """Emit an exit event with process details and duration."""
-    observation.emit(
-        "exit",
-        _EventDetails(
-            pid=pid,
-            exit_code=exit_code,
-            duration_s=max(0.0, exited_at - started_at),
-        ),
-    )
-
-
-@dc.dataclass(frozen=True, slots=True)
-class _TimeoutContext:
-    """Context information for a timeout exception."""
-
-    cmd_argv: tuple[str, ...]
-    timeout: float
-    stdout: str | None
-    stderr: str | None
-
-
 @dc.dataclass(frozen=True, slots=True)
 class _ExecutionTracking:
     """Hook and task tracking for command execution."""
@@ -445,138 +231,29 @@ class _ExecutionTracking:
     pending_tasks: list[asyncio.Task[None]]
 
 
-def _raise_timeout_expired(
-    timeout_ctx: _TimeoutContext,
-    exc: BaseException,
-) -> typ.NoReturn:
-    """Raise TimeoutExpired with captured output and chain the original exception."""
-    raise TimeoutExpired(
-        cmd=timeout_ctx.cmd_argv,
-        timeout=timeout_ctx.timeout,
-        output=timeout_ctx.stdout,
-        stderr=timeout_ctx.stderr,
-    ) from exc
-
-
 @dc.dataclass(frozen=True, slots=True)
-class _SubprocessTimeoutContext:
-    """Context for handling subprocess timeout exceptions."""
+class _IOBehaviour:
+    """Runtime I/O behaviour flags for command execution."""
 
-    execution: _SubprocessExecution
-    process: asyncio.subprocess.Process
-    started_at: float
-    stdout_text: str | None
-    stderr_text: str | None
+    capture: bool
+    echo: bool
 
 
-def _handle_subprocess_timeout(
-    ctx: _SubprocessTimeoutContext,
-    exc: TimeoutError | _SubprocessTimeoutError,
-) -> typ.NoReturn:
-    """Handle a subprocess timeout by emitting exit event and raising TimeoutExpired."""
-    if isinstance(exc, _SubprocessTimeoutError):
-        timeout = exc.timeout
-        exited_at = exc.exited_at
-        stdout = exc.stdout
-        stderr = exc.stderr
-    else:
-        timeout = ctx.execution.timeout
-        if timeout is None:
-            msg = "TimeoutError without a configured timeout"
-            raise RuntimeError(msg) from exc
-        exited_at = time.perf_counter()
-        stdout = ctx.stdout_text
-        stderr = ctx.stderr_text
-
-    exit_code = _get_exit_code(ctx.process)
-    _emit_exit_event(
-        ctx.execution.observation,
-        pid=ctx.process.pid,
-        exit_code=exit_code,
-        started_at=ctx.started_at,
-        exited_at=exited_at,
-    )
-    _raise_timeout_expired(
-        _TimeoutContext(
-            cmd_argv=ctx.execution.cmd.argv_with_program,
-            timeout=timeout,
-            stdout=stdout,
-            stderr=stderr,
-        ),
-        exc,
-    )
-
-
-async def _execute_subprocess(execution: _SubprocessExecution) -> CommandResult:
-    process = await _spawn_subprocess(execution)
-    started_at = time.perf_counter()
-    pid = process.pid
-    execution.observation.emit("start", _EventDetails(pid=pid))
-
-    stdout_text: str | None = None
-    stderr_text: str | None = None
-    try:
-        if not execution.capture and not execution.echo:
-            exit_code, exited_at = await _wait_for_exit_code(
-                process,
-                execution.ctx,
-                timeout=execution.timeout,
-            )
-        else:
-            (
-                exit_code,
-                exited_at,
-                stdout_text,
-                stderr_text,
-            ) = await _run_subprocess_with_streams(
-                process,
-                execution,
-                pid=pid,
-                timeout=execution.timeout,
-            )
-    except (TimeoutError, _SubprocessTimeoutError) as exc:
-        _handle_subprocess_timeout(
-            _SubprocessTimeoutContext(
-                execution=execution,
-                process=process,
-                started_at=started_at,
-                stdout_text=stdout_text,
-                stderr_text=stderr_text,
-            ),
-            exc,
-        )
-
-    _emit_exit_event(
-        execution.observation,
-        pid=pid,
-        exit_code=exit_code,
-        started_at=started_at,
-        exited_at=exited_at,
-    )
-
-    return CommandResult(
-        program=execution.cmd.program,
-        argv=execution.cmd.argv,
-        exit_code=exit_code,
-        pid=process.pid if process.pid is not None else -1,
-        stdout=stdout_text,
-        stderr=stderr_text,
-    )
-
-
-def _prepare_execution_observation(  # noqa: PLR0913
+def _prepare_execution_observation(
     cmd: SafeCmd,
     context: ExecutionContext,
     tracking: _ExecutionTracking,
-    *,
-    capture: bool,
-    echo: bool,
+    io_behaviour: _IOBehaviour,
 ) -> _StageObservation:
     """Prepare the observation context for command execution."""
     cwd = Path(context.cwd) if context.cwd is not None else None
     env_overlay = _freeze_str_mapping(context.env)
     tags = _merge_tags(
-        {"project": cmd.project.name, "capture": capture, "echo": echo},
+        {
+            "project": cmd.project.name,
+            "capture": io_behaviour.capture,
+            "echo": io_behaviour.echo,
+        },
         context.tags,
     )
     return _StageObservation(
@@ -652,12 +329,12 @@ class SafeCmd:
             execution_hooks=_run_before_hooks(self),
             pending_tasks=[],
         )
+        io_behaviour = _IOBehaviour(capture=capture, echo=echo)
         observation = _prepare_execution_observation(
             self,
             ctx,
             tracking,
-            capture=capture,
-            echo=echo,
+            io_behaviour,
         )
 
         observation.emit("plan", _EventDetails(pid=None))
